Brush applied and tested

+++ src/correlation.cpp
@@ -166,7 +166,7 @@ void pixel_filter(double rsltPixel[3], int x, int y, const unsigned char* origIm
 	// Added by Yingchuan
 	for (int rgb = 0; rgb < 3; rgb++)
 	{
-		int crossCorr = 0;
+		double crossCorr = 0;
 		for (int u = -knlWidth/2; u < knlWidth/2; u++)
 			for (int v = -knlHeight/2; v < knlWidth/2; v++)
 			{
